0.17.1: the instruments correct themselves
Both fixes were found BY the instruments on their first prod runs:

- DLA accumulated in bfloat16, and per-layer contributions of ±100
  cancelling to single digits turned its quantization into a ±3
  additivity residual — reported by the block's own honesty number.
  float32 now; the math-level test tightens from 5% to 1e-3. The
  additivity comparison also moves to pre-softcap space (gemma4 caps
  final logits; a linear decomposition cannot match a capped truth).
- patch/trace measured probability recovery, which is invisible for
  any target the clean prompt gives no mass to (explicit ' Paris'
  measured exactly nothing). metric: 'logprob' (new default)
  registers recovery at any mass; 'prob' remains for step-09 parity.

+++ mechbench_compute/attribution.py
@@ -260,7 +260,11 @@ def logit_attrs(
             )
         scale = float(np.asarray(ln_scale, dtype=np.float64).reshape(-1)[position])
         flat = (flat / scale) * _final_norm_gain(model)
-    v = mx.array(flat, dtype=mx.bfloat16)
+    # float32, deliberately: DLA components CANCEL (per-layer terms of
+    # ±100 summing to single digits), and bfloat16's ~0.5-at-100
+    # quantization accumulated a residual of ±3 across 36 components —
+    # caught by the block's own additivity check on its first prod run.
+    v = mx.array(flat, dtype=mx.float32)
 
     # Project through the unembed without applying the final norm — DLA
     # reads per-component contributions, and the final norm would mix
